CW2/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
pyodbc==4.0.39
requests==2.31.0
pydantic==1.10.7